refactor(LandscapePresetConfiguration): rename folder to landscape-preset-configuration
