+++ src/main/webapp/app/module/primary/landscape-preset-configuration/LandscapePresetConfiguration.component.ts
