+++ src/main/webapp/app/module/primary/landscape-preset-configuration/index.ts
